+++ superset-frontend/src/cccs-viz/plugins/plugin-chart-ag-grid/src/renderers/TimestampWithoutTimezoneValueRenderer.tsx
@@ -4,9 +4,8 @@ import moment from 'moment';
 
 const TimestampWithoutTimezoneValueRenderer: React.FC<{
   [index: string]: any;
-}> = React.memo(
-  // Since there's no timezone, we manually add in the UTC timezone
-  data => <>{moment(data.value).format('YYYY-MM-DD HH:mm:ss.SS UTC')}</>,
-);
+}> = React.memo(data => (
+  <>{moment(data.value).format('YYYY-MM-DD HH:mm:ss.SS')}</>
+));
 
 export default TimestampWithoutTimezoneValueRenderer;
